delete old lighting system
