+++ src/core/action.cpp
@@ -7,7 +7,6 @@
 #include "core/component/components.h"
 #include "core/factory/factory.h"
 #include "core/system/render.h"
-#include "core/system/lightning.h"
 #include "core/system/input/input.h"
 #include "core/system/render.h"
 
@@ -41,20 +40,20 @@ void createScene()
     EntityFactory::createPlane(Vector3{ 0.f, 1.f, 0.f }, 0.f);
 
     // New light entities
-    const Shader& shader = RenderSystem::getSystem().getGeometryShader();
-    auto const lights = {
+    //const Shader& shader = RenderSystem::getSystem().getGeometryShader();
+    /*auto const lights = {
         EntityFactory::createLight(shader, LIGHT_DIRECTIONAL, Vector3{ 20.f, 70.f, 0.f }, Vector3Zero(), WHITE),
         EntityFactory::createLight(shader, LIGHT_SPOT, Vector3{ -2.f, 2.f, -2.f }, Vector3Zero(), WHITE, CAMERA_PERSPECTIVE),
         EntityFactory::createLight(shader, LIGHT_POINT, Vector3{ 2.f, 1.f, 2.f }, Vector3Zero(), YELLOW, CAMERA_ORTHOGRAPHIC, 0.f),
     };
-    LightningSystem::getSystem().setActiveLightEntity(lights.begin()[0]);
+    LightningSystem::getSystem().setActiveLightEntity(lights.begin()[0]);*/
 
     constexpr float cutoff  = 0.7f;
     constexpr float radius  = 10.f;
     constexpr float softness= 0.65f;
-    EntityFactory::create(WHITE , Position{ 20.f, 70.f, 0.f }    , ShadowCaster{ CAMERA_ORTHOGRAPHIC }  , LookAt{}  , DirectionalLight{});
+    /*EntityFactory::create(WHITE , Position{ 20.f, 70.f, 0.f }   , ShadowCaster{ CAMERA_ORTHOGRAPHIC }   , LookAt{}  , DirectionalLight{});
     EntityFactory::create(YELLOW, Position{ 2.f, 1.f, 2.f }     , ShadowCaster{ CAMERA_PERSPECTIVE }    , PointLight{ radius, softness });
-    EntityFactory::create(WHITE , Position{ -2.f, 2.f, -2.f }   , ShadowCaster{ CAMERA_PERSPECTIVE }    , LookAt{}  , SpotLight { radius, softness, cutoff });
+    EntityFactory::create(WHITE , Position{ -2.f, 2.f, -2.f }   , ShadowCaster{ CAMERA_PERSPECTIVE }    , LookAt{}  , SpotLight { radius, softness, cutoff });*/
 }
 
 void throwBallFromCamera()
