+++ src/core/component/components.h
@@ -1,6 +1,6 @@
 #pragma once
+
 #include "raylib.h"
-#include "graphics/lights.h"
 
 struct TransformComponent
 {
@@ -28,27 +28,6 @@ struct ClickedEntityTag
 { 
 };
 
-struct LightComponent
-{
-    Light light;
-    Camera caster;
-};
-
-struct Inactive     {};
+struct Inactive {};
 struct Position : Vector3 {};
 struct LookAt   : Vector3 {};
-struct DirectionalLight {};
-
-struct ShadowCaster {
-    CameraProjection projection;
-};
-
-
-struct PointLight {
-    float lightRadius;
-    float softness;
-};
-
-struct SpotLight : PointLight {
-    float cutoff;
-};
