+++ src/graphics/light/components.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#include "raylib.h"
+
+struct DirectionalLight {};
+
+struct ShadowCaster {
+    CameraProjection projection;
+};
+
+
+struct PointLight {
+    float lightRadius;
+    float softness;
+};
+
+struct SpotLight : PointLight {
+    float cutoff;
+};
